requirements/test.txt
pytest==2.6.4
pytest-django==2.8.0
pytest-pythonpath==0.6
pytest-xdist
git+http://github.com/mverteuil/pytest-ipdb.git
factory-boy==2.4.1

xtraceback==0.4.0-rc1
pylint==1.2.1
clonedigger==1.1.0
autopep8==1.0.4
pep8ify==0.0.12
pep8==1.5.7

#old requirements for nose
#django-nose==1.2
#nose==1.3.4
#nose-cov==1.6
#nose-progressive==1.5.1
